requirements.txt
aiohappyeyeballs  
aiohttp  
aiosignal  
annotated-types  
anyio  
appnope  
argon2-cffi  
argon2-cffi-bindings  
arrow  
asttokens  
async-timeout  
attrs  
backcall  
beautifulsoup4  
bleach  
blinker  
build  
certifi  
cffi  
charset-normalizer  
click  
comm  
debugpy  
decorator  
defusedxml  
Deprecated  
dill  
distlib  
distro  
entrypoints  
exceptiongroup  
executing  
faiss-cpu  
fastjsonschema  
filelock  
Flask  
Flask-Limiter  
fqdn  
frozenlist  
fsspec  
fuzzywuzzy  
h11  
httpcore  
httpx  
huggingface-hub  
idna  
importlib-metadata  
importlib_resources  
ipykernel  
ipython  
ipython-genutils  
isoduration  
itsdangerous  
jedi  
Jinja2  
jiter  
joblib  
jsonpointer  
jsonschema  
jupyter-events  
jupyter_client  
jupyter_core  
jupyter_server  
jupyter_server_terminals  
jupyterlab-pygments  
limits  
markdown-it-py  
MarkupSafe  
matplotlib-inline  
mdurl  
mistune  
mpmath  
multidict  
multiprocess  
nbclassic  
nbclient  
nbconvert  
nbformat  
nest-asyncio  
networkx  
nltk  
notebook  
notebook_shim  
numpy  
openai  
ordered-set  
packaging  
pandocfilters  
parso  
pexpect  
pickleshare  
Pillow  
pip-tools  
platformdirs  
prometheus-client  
prompt-toolkit  
propcache  
psutil  
psycopg2-binary  
ptyprocess  
pure-eval  
pycparser  
pydantic  
pydantic_core  
Pygments  
pyproject_hooks  
pyrsistent  
python-dateutil  
python-dotenv  
python-json-logger  
PyYAML  
pyzmq  
RapidFuzz  
redis  
regex  
requests  
rfc3339-validator  
rfc3986-validator  
rich  
safetensors  
scikit-learn  
scipy  
Send2Trash  
six  
sniffio  
soupsieve  
stack-data  
sympy  
tdqm  
terminado  
thefuzz  
threadpoolctl  
tinycss2  
tokenizers  
tomli  
torch  
torchaudio  
torchvision  
tornado  
tqdm  
traitlets  
transformers  
typing_extensions  
uri-template  
urllib3  
virtualenv  
wcwidth  
webcolors  
webencodings  
websocket-client  
Werkzeug  
wrapt  
yarl  
zipp